+++ dantro/plot_mngr.py
@@ -50,7 +50,7 @@ class PlotManager:
     """
 
     CREATORS = ALL_PCRS
-    DEFAULT_OUT_FSTRS = dict(date="%y%m%d-%H%M%S",
+    DEFAULT_OUT_FSTRS = dict(timestamp="%y%m%d-%H%M%S",
                              state_no="{no:0{digits:d}d}",
                              state="{name:}_{val:}",
                              state_name_replace_chars=[], # (".", "-")
@@ -65,7 +65,14 @@ class PlotManager:
                              plot_cfg_sweep="{name:}/sweep_cfg.yml",
                              )
 
-    def __init__(self, *, dm: DataManager, plots_cfg: Union[dict, str]=None, out_dir: Union[str, None]="{date:}/", out_fstrs: dict=None, creator_init_kwargs: Dict[str, dict]=None, default_creator: str=None, save_plot_cfg: bool=True, raise_exc: bool=False):
+    def __init__(self, *, dm: DataManager, plots_cfg: Union[dict, str]=None,
+                 out_dir: Union[str, None]="{timestamp:}/",
+                 out_fstrs: dict=None,
+                 creator_init_kwargs: Dict[str, dict]=None,
+                 default_creator: str=None,
+                 save_plot_cfg: bool=True,
+                 raise_exc: bool=False,
+                 cfg_exists_action: str='raise'):
         """Initialize the PlotManager
         
         Args:
@@ -76,18 +83,18 @@ class PlotManager:
             out_dir (Union[str, None], optional): If given, will use this
                 output directory as basis for the output path for each plot.
                 The path can be a format-string; it is evaluated upon call to
-                the plot command. Available keys: `date`, `name`, ...
+                the plot command. Available keys: `timestamp`, `name`, ...
                 For a relative path, this will be relative to the DataManager's
                 output directory. Absolute paths remain absolute.
                 If this argument evaluates to False, the DataManager's output
                 directory will be the output directory.
             out_fstrs (dict, optional): Format strings that define how the
                 output path is generated. The dict given here updates the
-                DEAFULT_OUT_FSTRS class variable which holds the default values
-                Keys: `date` (%-style), `path`, `sweep`, `state`, `plot_cfg`,
-                      `state`, `state_no`, `state_join_char`,
+                DEFAULT_OUT_FSTRS class variable which holds the default values
+                Keys: `timestamp` (%-style), `path`, `sweep`, `state`,
+                      `plot_cfg`, `state`, `state_no`, `state_join_char`,
                       `state_vector_join_char`
-                Available keys for `path`: `name`, `date`, `ext`
+                Available keys for `path`: `name`, `timestamp`, `ext`
                 Additionally, for `sweep`: `state_no`, `state_vector`, `state`
             creator_init_kwargs (Dict[str, dict], optional): If given, these
                 kwargs are passed to the initialization calls of the respective
@@ -99,6 +106,8 @@ class PlotManager:
             raise_exc (bool, optional): Whether to raise exceptions if there
                 are errors raised from the plot creator or errors in the plot
                 configuration. If False, the errors will only be logged.
+            cfg_exists_action (str, optional): Behaviour when a config file
+                already exists. Can be: skip, overwrite, raise.
         
         Raises:
             InvalidCreator: When an invalid default creator was chosen
@@ -115,6 +124,9 @@ class PlotManager:
         # Private or read-only
         self._dm = dm
         self._out_dir = out_dir
+
+        # Parameters to pass through as defaults to member functions
+        self._cfg_exists_action = cfg_exists_action
 
         # Handle plot config
         if isinstance(plots_cfg, str):
@@ -161,21 +173,24 @@ class PlotManager:
 
     def _parse_out_dir(self, fstr: str, *, name: str) -> str:
         """Evaluates the format string to create an output directory.
-
+        
         Note that the directories are _not_ created; this is outsourced to the
         plot creator such that it happens as late as possible.
         
         Args:
             fstr (str): The format string to evaluate and create a directory at
+            name (str): Name of the plot
+            timestamp (float, optional): Description
         
         Returns:
             str: The path of the created directory
         """
-        # Get date format string and evaluate
-        date_fstr = self._out_fstrs.get('date', "%y%m%d-%H%M%S")
-        date = time.strftime(date_fstr)
+        # Get date format string and current time and create a string
+        # TODO allow passing a timestamp?
+        timefstr = self._out_fstrs.get('timestamp', "%y%m%d-%H%M%S")
+        timestr = time.strftime(timefstr)
 
-        out_dir = fstr.format(date=date, name=name)
+        out_dir = fstr.format(timestamp=timestr, name=name)
 
         # Make sure it is absolute
         if not os.path.isabs(out_dir):
@@ -185,7 +200,10 @@ class PlotManager:
         # Return the full path
         return out_dir
 
-    def _parse_out_path(self, creator: BasePlotCreator, *, name: str, out_dir: str, file_ext: str=None, state_no: int=None, state_no_max: int=None, state_vector: Tuple[int]=None, dims: dict=None) -> str:
+    def _parse_out_path(self, creator: BasePlotCreator, *, name: str,
+                        out_dir: str, file_ext: str=None, state_no: int=None,
+                        state_no_max: int=None, state_vector: Tuple[int]=None,
+                        dims: dict=None) -> str:
         """Given a creator and (optionally) parameter sweep information, a full
         and absolute output path is generated, including the file extension.
         
@@ -209,7 +227,8 @@ class PlotManager:
         Returns:
             str: The fully parsed output path for this plot
         """
-        def parse_state_pair(name: str, dim: ParamDim, *, fstrs: dict) -> Tuple[str]:
+        def parse_state_pair(name: str, dim: ParamDim, *,
+                             fstrs: dict) -> Tuple[str]:
             """Helper method to create a state pair"""
             # Parse the name
             for search, replace in fstrs['state_name_replace_chars']:
@@ -227,7 +246,7 @@ class PlotManager:
         fstrs = self.out_fstrs
         
         # Evaluate the keys available for both cases
-        keys = dict(date=time.strftime(fstrs['date']), name=name)
+        keys = dict(timestamp=time.strftime(fstrs['timestamp']), name=name)
 
         # Parse file extension and ensure it starts with a dot
         ext = file_ext if file_ext else creator.get_ext()
@@ -298,7 +317,8 @@ class PlotManager:
 
         return rv
 
-    def _store_plot_info(self, name: str, *, plot_cfg: dict, creator_name: str, save: bool, target_dir: str, **info):
+    def _store_plot_info(self, name: str, *, plot_cfg: dict, creator_name: str,
+                         save: bool, target_dir: str, **info):
         """Stores all plot information in the plot_info list and, if `save` is
         set, also saves it using the _save_plot_cfg method.
         """
@@ -319,8 +339,10 @@ class PlotManager:
         # Append to the plot_info list
         self._plot_info.append(entry)
 
-    def _save_plot_cfg(self, cfg: dict, *, name: str, creator_name: str, target_dir: str, is_sweep: bool=False) -> str:
-        """Saves the given configuration under the top-level entry `name` t, is_sweep=Trueo
+    def _save_plot_cfg(self, cfg: dict, *, name: str, creator_name: str,
+                       target_dir: str, exists_action: str=None,
+                       is_sweep: bool=False) -> str:
+        """Saves the given configuration under the top-level entry `name` to
         a yaml file.
         
         Args:
@@ -328,16 +350,30 @@ class PlotManager:
             name (str): The name of the plot
             creator_name (str): The name of the creator
             target_dir (str): The directory path to store the file in
+            exists_action (str, optional): What to do if a plot configuration
+                already exists. Can be: overwrite, skip, append, raise.
+                If None, uses the value 'cfg_exists_action' given during
+                initialization of the PlotManager.
+            is_sweep (bool, optional): Set if the configuration refers to a
+                plot in sweep mode, for which a different format string is used
         
         Returns:
             str: The path the config was saved at (mainly used for testing)
+        
+        Raises:
+            ValueError: For invalid `exists_action` argument
         """
+        # Resolve default arguments
+        if exists_action is None:
+            exists_action = self._cfg_exists_action
+
         # Build the dict that is to be saved
         d = dict()
         d[name] = copy.deepcopy(cfg)
 
         if not isinstance(cfg, ParamSpace):
             d[name]['creator'] = creator_name
+
         else:
             # FIXME hacky, should not use the internal API!
             d[name]._dict['creator'] = creator_name
@@ -345,13 +381,39 @@ class PlotManager:
         # Generate the filename
         if not is_sweep:
             fname = self.out_fstrs['plot_cfg'].format(name=name)
+
         else:
             fname = self.out_fstrs['plot_cfg_sweep'].format(name=name)
         save_path = os.path.join(target_dir, fname)
         
-        # And save
-        write_yml(d, path=save_path)
-        log.debug("Saved plot configuration for '%s' to: %s", name, save_path)
+        # Try to write
+        try:
+            write_yml(d, path=save_path, mode='x')
+
+        except FileExistsError as err:
+            log.debug("Config file already exists at %s!", save_path)
+            
+            if exists_action == 'raise':
+                raise
+
+            elif exists_action == 'skip':
+                log.debug("Skipping ...")
+
+            elif exists_action == 'append':
+                log.debug("Appending ...")
+                write_yml(d, path=save_path, mode='a')
+
+            elif exists_action == 'overwrite':
+                log.debug("Overwriting ...")
+                write_yml(d, path=save_path, mode='w')
+
+            else:
+                raise ValueError("Invalid value '{}' for argument "
+                                 "`exists_action`!".format(exists_action))
+
+        else:
+            log.debug("Saved plot configuration for '%s' to: %s",
+                      name, save_path)
 
         return save_path